### 2022-07-26 #4610 @editorialbot
[PRE REVIEW]: quadcleanR: An R Package for the Cleanup and Visualization of Quadrat Data
**Submitting author:** @DominiqueMaucieri (<a href="http://orcid.org/0000-0003-1849-2472">Dominique G. Maucieri</a>)
**Repository:** https://github.com/DominiqueMaucieri/quadcleanR
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
