### 2022-07-27 #4610 @Kevin-Mattheus-Moerman
@DominiqueMaucieri I am an AEiC with JOSS and here to process the first steps and to help assign an editor if this work is deemed in scope.

- [ ] Since you are the sole author, you can remove the statement (and associated YML) that authors contributed equally.
- [ ] Please check the above potentially missing DOIs. You can get our bot to check them again by calling `@editorialbot check references`.
