### 2022-07-27 28c5316
Adding DOI's, JOSS badge and fixing author contribution status
