+++ manuscript/paper.bib
@@ -6,6 +6,7 @@
 	Title = {Towards Automated Annotation of Benthic Survey Images: Variability of Human Experts and Operational Modes of Automation},
 	Volume = 10,
 	Year = 2015,
+	doi: {10.1371/journal.pone.0130312},
         url = {https://doi.org/10.1371/journal.pone.0130312}
 }
 
@@ -17,6 +18,7 @@
 	Title = {A rapid ecological assessment (REA) quantitative survey method for benthic algae using photoquadrats with SCUBA},
 	Volume = 58,
 	Year = 2004,
+	doi: {10.1353/psc.2004.0021},
         url = {https://doi.org/10.1353/psc.2004.0021}
 }
 
@@ -28,6 +30,7 @@
 	Title = {Estimating the abundance of benthic invertebrates: a comparison of procedures and variability between observers},
 	Volume = 138,
 	Year = 1996,
+	doi: {10.3354/meps138093},
         url = {https://doi.org/10.3354/meps138093}
 }
 
@@ -40,15 +43,17 @@
 	Title = {Whittaker's plant diversity sampling method},
 	Volume = 33,
 	Year = 1984,
+	doi: {10.1080/0021213X.1984.10676984},
         url = {https://doi.org/10.1080/0021213X.1984.10676984}
 }
 
 @book{williams1964,
-  Title={Patterns in the balance of nature and related problems of quantitative ecology.},
+  Title={Patterns in the balance of nature and related problems of quantitative ecology},
   Author={Williams, Carrington Bonsor},
   Year = {1964},
   Publisher={Academic Press, London},
-  Pages = {324}
+  Pages = {324},
+	doi: {10.2307/2257968}
 }
 
 
@@ -57,7 +62,8 @@
   author={Clements, Frederic Edward},
   number={242},
   year={1916},
-  publisher={Carnegie Institution of Washington}
+  publisher={Carnegie Institution of Washington},
+	doi: {10.5962/bhl.title.56234}
 }
 
 
@@ -69,6 +75,7 @@
   number={7542},
   pages={171--180},
   year={2015},
+	doi: {10.1038/nature14258},
         url = {https://doi.org/10.1038/nature14258}
 }
 
@@ -80,6 +87,7 @@
   number={7656},
   pages={65--72},
   year={2017},
+	doi: {10.1038/nature22899},
         url = {https://doi.org/10.1038/nature22899}
 }
 
@@ -101,6 +109,7 @@
   number={8},
   pages={1233--1244},
   year={2018},
+	doi: {10.1111/ecog.03148},
       url = {https://doi.org/10.1111/ecog.03148}
 }
 
@@ -110,5 +119,6 @@
   journal={Journal of Biogeography},
   pages={421--428},
   year={1993},
+	doi: {10.2307/2845590},
       url = {https://doi.org/10.2307/2845590}
 }
