+++ manuscript/paper.bib
@@ -42,9 +42,7 @@
 	Pages = {41--46},
 	Title = {Whittaker's plant diversity sampling method},
 	Volume = 33,
-	Year = 1984,
-	doi = {10.1080/0021213X.1984.10676984},
-        url = {https://doi.org/10.1080/0021213X.1984.10676984}
+	Year = 1984
 }
 
 @book{williams1964,
