+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   cleaning_cropping: cleaning_cropping.html
   cleaning_simple: cleaning_simple.html
   cropping: cropping.html
-last_built: 2022-07-27T20:55Z
+last_built: 2022-07-27T21:00Z
 urls:
   reference: https://dominiquemaucieri.github.io/quadcleanR/reference
   article: https://dominiquemaucieri.github.io/quadcleanR/articles
