### 2022-08-11 #4610 @fboehm
@jhollist @SteveViss - would you be willing to review this submission for JOSS? Our reviews are checklist-driven. Thanks!
