### 2022-08-11 #4610 @ViniciusBRodrigues
> @ha0ye @juanvillada @ViniciusBRodrigues @shaunpwilkinson - would you be willing to review this submission for JOSS? Our reviews are checklist-driven. Please feel free to ask me any questions if you're unfamiliar with our review process. Thanks!
Hi there. I'm glad to help.
