### 2022-08-18 #4674 @diazrenata
## Review summary

The primary functionality of the `quadcleanR` package is the `crop_area` function, which standardizes the sampling density and effort of data collected from quadrats of different dimensions to allow for robust comparisons between different-sized quadrats. The package also includes a number of data wrangling utility functions and a shiny app for basic exploratory plotting.

I can see this being a useful contribution, given that many ecologists have to confront quadrats of nonstandard area. The functions work as described, the function documentation is sufficient, and the cheatsheet and shiny app help with user onboarding. I do have some suggestions that would help clarify how this package adds to the wider ecosystem of available tools, and to improve the package vignettes as aids to a general user population.

## Specific suggestions

### Writing

This draft of the manuscript does not contain a current state of the field section identifying how this package complements existing tools.

This would be especially helpful, e.g. for the cleaning functions (a number of which are similar to functionality available in tidyverse, or operations I would be just as likely to code up from scratch specific to a given dataset) and the shiny app (which renders basic exploratory plots that seem similar to what users are probably doing on their own).

(For example - a general user might find it simpler to use `dplyr::select` or `dplyr::filter` rather than learn the syntax for `quadcleanR::keep_rm`, or the `dplyr::rename` family rather than create an extra dataframe and learn `quadcleanR::change_names`).

### Documentation

I suggest adding structure to the vignettes to tighten the focus on the added-value of the functions provided in this package. For example, the Simple Cleaning Quadrat Data vignette is a detailed narrative walk through cleaning a specific dataset, including a mix of `quadcleanR` and other package functions woven together. That would make it a great methods supplement to a manuscript, but it's not so easy for a user to read it and quickly understand the `quadcleanR` functions. There is also a fair amount of repeated text between the different vignettes, which makes it harder for a user to extract the important information from each one.

You don't have to necessarily do it this way, but one suggestion would be to create a more detailed outline of the topics/functions highlighted in each vignette and structuring each section around demonstrating  that group of functions.

It may also be helpful to use a lightly preprocessed version of the datasets used in the vignette, to get some of the additional data cleaning out of the way so you can focus the vignette specifically on `quadcleanR` functions.

### Shiny app

The demo version of the shiny app (https://dominiquemaucieri.shinyapps.io/example/) is fun but has a lot of variables that aren't explained. With so many options, it's hard for someone who isn't familiar with this dataset to use the dashboard to generate informative plots. I think it'd be a more useful demo for a naive user if there were just a few options for faceting/grouping variables, and these were either described or given very intuitive names.

### Substantive scholarly effort

I leave this up to the editor, and anything clarified in a "state-of-the-field" section of the manuscript, but from my perspective the primary unique contribution for this package is the `crop_area` function. There are a number of cleaning functions that are excellent utilities, but for me are in a grey area between useful scripts for streamlining a specific data cleaning workflow and general-use utilities that substantially improve on functionality already available. Again, I'm happy to see this clarified with a little more conversation.

### Code

Finally, I just want to note that I've run and tested the functions as they are written, but I haven't done extensive stress-testing to see if I can break them with manufactured use cases. I notice that in the source code there isn't a lot of error handling built in. I think this is reasonable for this package because most of the functions are pretty small and lightweight, but in general I would encourage more defensive programming in a package intended for widespread use.
