### 2022-08-31 c58d758
updating vignette names and creating function_examples vignette
