+++ data-raw/add_coral_data.R
@@ -30,8 +30,8 @@ environmental_data <- read.csv(here::here("data-raw", "environmental_data.csv"))
 usethis::use_data(environmental_data, overwrite = TRUE,
                   compress = "xz")
 
-simple_cleaned <- read.csv(here::here("data-raw", "simple_cleaned.csv"))
+final_cleaned <- read.csv(here::here("data-raw", "final_cleaned.csv"))
 
-usethis::use_data(simple_cleaned, overwrite = TRUE,
+usethis::use_data(final_cleaned, overwrite = TRUE,
                   compress = "xz")
 
