+++ vignettes/cleaning_coral.Rmd
@@ -1,19 +1,19 @@
 ---
-title: "Simple Cleaning Quadrat Data"
+title: "Cleaning Quadrat Data from CoralNet"
 description: >
   A simple walk through of how to clean up quadrat data using the `quadcleanR`
   package. This will go through all the functions except the `crop_area()`
   function, as well as visualizing the resulting data.
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Simple Cleaning Quadrat Data}
+  %\VignetteIndexEntry{Cleaning Quadrat Data from CoralNet}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
 # Introduction
 
-This vignette will walk you through cleaning up quadrat data to produce an easy to analyze data frame.
+This vignette will walk you through cleaning up quadrat data from [CoralNet](https://coralnet.ucsd.edu/) to produce an easy to analyze data frame.
 
 A little about the data:
 
