+++ vignettes/cleaning_coral.Rmd
@@ -227,7 +227,7 @@ The final things I will add to this data to get it in shape for analysis is a fi
 
 B_LQuad_timeblock <- categorize(B_LQuad_enviro, column = "Field.Season", values = unique(B_LQuad_enviro$Field.Season), name = "TimeBlock", binary = FALSE, exact = TRUE, categories = c(rep("Before", times = 4), rep("During", times = 3), rep("After", times = 4)))
 
-simple_cleaned <- keep_rm(B_LQuad_timeblock, values = "Soft_coral", select = "row", colname = "functional_group")
+final_cleaned <- keep_rm(B_LQuad_timeblock, values = "Soft_coral", select = "row", colname = "functional_group")
 
 ```
 
@@ -238,14 +238,14 @@ This data has now been sufficiently cleaned and can be used for many different a
 
 ```{r sample sizes, results = 'hide'}
 
-sample_size(simple_cleaned, dim_1 = "Site", dim_2 = "Field.Season", count = "Quadrat")
+sample_size(final_cleaned, dim_1 = "Site", dim_2 = "Field.Season", count = "Quadrat")
 
 ```
 
 
 ```{r sample sizes cleaner md output, message = FALSE, echo = FALSE, results='asis'}
 
-knitr::kable(sample_size(simple_cleaned, dim_1 = "Site", dim_2 = "Field.Season", count = "Quadrat"), align = 'c')%>%
+knitr::kable(sample_size(final_cleaned, dim_1 = "Site", dim_2 = "Field.Season", count = "Quadrat"), align = 'c')%>%
   kable_styling("striped", full_width = F) %>% 
   scroll_box(width = "100%")
 
@@ -264,7 +264,7 @@ The data used in this example is this data set we have just cleaned. However thi
 
 ```{r shiny app, eval= FALSE}
 
-visualize_app(data = simple_cleaned, xaxis = colnames(simple_cleaned[,1:13]), yaxis = "prop_cover")
+visualize_app(data = final_cleaned, xaxis = colnames(final_cleaned[,1:13]), yaxis = "prop_cover")
 
 ```
 
