+++ _pkgdown.yml
@@ -40,6 +40,7 @@ reference:
 
 - title: Data
   contents:
+  - corals
   - coral_labelset
   - environmental_data
   - final_cleaned
