+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   cleaning_cropping_coral: cleaning_cropping_coral.html
   cropping: cropping.html
   function_examples: function_examples.html
-last_built: 2022-08-31T18:48Z
+last_built: 2022-08-31T21:35Z
 urls:
   reference: https://dominiquemaucieri.github.io/quadcleanR/reference
   article: https://dominiquemaucieri.github.io/quadcleanR/articles
