### 2022-08-31 da2b925
creating function_examples vignette
