+++ vignettes/function_examples.Rmd
@@ -1,10 +1,10 @@
 ---
 title: "Simple Examples of Functions"
 description: >
-  Examples of how to use the `quadcleanR` package functions for data cleaning. 
-  This vignette will go through all the functions except the `crop_area()`
-  function, which is explained in further details in the Why to Crop Quadrats by Area
-  vignette.
+  Examples of how to use the `quadcleanR` package functions for data 
+  cleaning and processing. This vignette will go through all the functions 
+  except the `crop_area()` function, which is explained in further details 
+  in the Why to Crop Quadrats by Area vignette.
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Simple Examples of Functions}
@@ -12,12 +12,158 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r setup}
-library(quadcleanR)
-```
-
-
 # Introduction
 
 This vignette will walk you through the different functions for cleaning quadrat data within the 
 'quadcleanR' package. 
+
+```{r setup, warning=FALSE, message=FALSE}
+
+library(quadcleanR)
+
+data("corals")
+
+```
+
+
+# Cleaning Functions
+
+There are 5 cleaning functions within quadcleanR.
+
+## 1. change_names()
+
+Using a new data frame of labels, change column names in one function. Helpful if column names are shorthands or contain spaces and characters that are not supported in column names in R.
+
+```{r change_names, warning=FALSE, message=FALSE}
+
+data("coral_labelset")
+
+corals_change_names <- change_names(corals, coral_labelset, from = "short_name", to = "full_name")
+
+```
+
+
+## 2. change_values()
+
+Using two vectors, change the values in one column to a new set of values. Helpful if you need to change many values at once, like updating changes to site names or taxonomy.
+
+```{r change_values, warning=FALSE, message=FALSE}
+
+corals_change_values <- change_values(corals_change_names, "Field.Season", c("KI2015a", "KI2016a"), c("KI2015", "KI2016"))
+
+```
+
+
+## 3. keep_rm()
+Using a character, or part of character select rows or columns of the data frame to either keep or remove. A more customizable way to subset your data as you can keep or remove based on partial matches, or cells containing select characters.
+
+```{r keep_rm, warning=FALSE, message=FALSE}
+
+corals_keep_rm <- keep_rm(corals_change_values, "DEEP" , select = "row", keep = FALSE, drop_levels = TRUE, exact = FALSE, "Site")
+
+```
+
+
+## 4. rm_chr()
+Parts of characters can be removed based on a vector of removal characters. When workingwith images, this can be helpful to remove extra characters from image IDs, or anywhere else where you want to remove specific characters from your data. 
+
+
+```{r rm_chr, warning=FALSE, message=FALSE}
+
+corals_rm_chr <- rm_chr(corals_keep_rm, rm = c(".jpg", ".jpeg"), full_selection = FALSE, cols = "Quadrat")
+
+```
+
+
+## 5. sum_cols()
+Select columns and attach a vector of their new names, then columns with matching names will have each row summed. This is helpful to simplify your data quickly, like simplifying at a higher taxonomic group.
+
+```{r sum_cols, warning=FALSE, message=FALSE}
+
+corals_sum_cols <- sum_cols(corals_rm_chr, from = c("Acropora_corymbose", "Acropora_digitate", "Acropora_tabulate"), to = c("Acropora_spp", "Acropora_spp", "Acropora_spp"))
+
+```
+
+
+# Adding Functions
+
+There are 2 adding functions within quadcleanR.
+
+## 1. add_data()
[…]
