### 2022-09-01 1be606b
adding ability to deal with NAs and NaNs to cover_calc() and usable_obs()
