+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   cleaning_cropping_coral: cleaning_cropping_coral.html
   cropping: cropping.html
   function_examples: function_examples.html
-last_built: 2022-08-31T21:44Z
+last_built: 2022-09-01T23:54Z
 urls:
   reference: https://dominiquemaucieri.github.io/quadcleanR/reference
   article: https://dominiquemaucieri.github.io/quadcleanR/articles
