### 2022-09-06 6fb1098
updating news and editing manuscript to address reviewer comments
