+++ manuscript/paper.md
@@ -20,15 +20,15 @@ affiliations:
 
 # Summary
 
-Biodiversity is a crucial component of ecosystems, such that the monitoring of biodiversity is a priority for effective conservation programs. For communities of sessile organisms, it is common to implement quadrats (a frame isolating a known area) as a tool to quantify the biodiversity of an ecosystem. With technological advancements like photo quadrats and machine learning, there is much potential for the collection of quadrat data from different monitoring programs and for studies to conduct larger scale analyses. However, for quadrat data from different methodologies to be compatible, data must be standardized to the same area. The [quadcleanR](https://dominiquemaucieri.com/quadcleanR/) R package was created to assist in standardizing quadrat areas while maintaining the spatial relationships between identified organisms and substrates as well as assist in the cleanup and visualization of quadrat data. This package aims to facilitate comparability between quadrat data from different sources and streamline data cleanup.
+Biodiversity is a crucial component of ecosystems, such that the monitoring of biodiversity is a priority for effective conservation programs. For communities of sessile organisms, it is common to implement quadrats (a frame isolating a known area) to quantify the ecosystem’s biodiversity. Technological advancements like photo quadrats and machine learning provide potential for the collection of quadrat data from different monitoring programs and for studies to conduct larger scale analyses. However, for quadrat data from different methodologies to be compatible, data must be standardized to the same area. The [quadcleanR](https://dominiquemaucieri.com/quadcleanR/) R package was created to assist in standardizing quadrat areas while maintaining the spatial relationships between identified organisms and substrates as well as assist in the cleanup and visualization of quadrat data. This package aims to facilitate comparability between quadrat data from different sources and streamline data cleanup.
 
 # Statement of need
 
-Human actions are driving losses in global biodiversity [@LewisMaslin] making the study of drivers of biodiversity a priority for the conservation of ecosystems [@isbell2017]. For communities of sessile or slow-moving organisms, the use of quadrats allows for standardization of size and shape from which changes in community composition and structure can be recorded and monitored spatially and temporally [@clements1916plant]. Photos of quadrats increase the speed that data can be collected _in situ_, as the processing time can occur years later by trained researchers, instead of at the time of collection [@Preskitt:2004]. Nowadays, photo quadrats are processed using a series of computer-generated points (either a uniform grid or random placement) being superimposed onto the image and the resulting organism or substrate that falls under these points are identified [@Preskitt:2004]. Furthermore, technological advances in machine learning have supported the inclusion of deep learning in quadrat analyses [@Beijbom:2015] facilitating automated identification of organisms and substrates and increasing the speed that quadrats can be annotated with a lower associated cost [@Beijbom:2015]. However, these technological advances introduce challenges associated with quadrats collected under different methodologies. 
[…]
