+++ NEWS.md
@@ -1,7 +1,19 @@
+# quadcleanR 1.0.0.9000
+
+* Manuscript submitted to the Journal of Open Source Software with reviewer comments leading to package updates.
+
+* New vignette created "Simple Examples of Functions" to highlight the uses of each package in a simple manner that is easy to follow.
+
+* New data frame "corals" to accompany the "Simple Examples of Functions" vignette.
+
+* Renaming `print_max` with `above_cutoff` within the `usable_obs()` function to ensure clarity. Vignettes and cheatsheet were updated to reflect this change.
+
+* `usable_obs()` and `cover_calc()` were updated to be able to better deal with 0s, NAs and NaNs. 
+
 # quadcleanR 1.0.0
 
-First public release.
+* First public release.
 
 # quadcleanR 0.0.0.9000
 
-Initial Package development.
+* Initial Package development.
