+++ manuscript/paper.md
@@ -26,7 +26,7 @@ Biodiversity is a crucial component of ecosystems, such that the monitoring of b
 
 Human actions are driving losses in global biodiversity [@LewisMaslin] making the study of drivers of biodiversity a conservation priority [@isbell2017]. For communities of sessile or slow-moving organisms, quadrats allow for standardization of size and shape from which changes in community composition and structure can be recorded and monitored spatially and temporally [@clements1916plant]. Photos of quadrats increase the speed that data can be collected _in situ_, as the processing time can occur later by trained researchers, instead of at the time of collection [@Preskitt:2004]. Nowadays, photo quadrats are processed using a series of computer-generated points (either a uniform grid or random placement) being superimposed onto the image and the resulting organism or substrate that falls under these points are identified [@Preskitt:2004]. Technological advances in machine learning have supported the inclusion of deep learning in quadrat analyses [@Beijbom:2015] facilitating automated identification of organisms and substrates and increasing the speed that quadrats can be annotated with a lower associated cost [@Beijbom:2015]. However, these technological advances introduce challenges associated with quadrats collected under different methodologies. 
 
-For collaborative quadrat analyses, it is crucial that quadrat data are being properly compared to account for different methodologies. Size of quadrats varies based on the research question, study site and target species [@Benedetti:1996]; however, biomass and diversity estimates are greatly influenced by the size of quadrats [@Shmida; @williams1964; @anderson1993effect]. Thus, issues can arise when studies examine many different quadrats from various locations, which have been sampled using different quadrat sizes. If quadrat samples are being cropped to the same size before species are identified, this will produce comparable sized quadrats to be used in analysis; however, if experts have identified the species in their quadrats prior to sharing data, researchers must ensure that methods taken to standardize the size of quadrats properly considers the spatial nature of quadrat data. Currently there are no standardized methods to account for the size of quadrats and as global studies become more accessible, there is a need for standardizing these methodologies. The [quadcleanR](https://dominiquemaucieri.com/quadcleanR/) R package provides tools to assist in standardizing quadrat areas and the cleanup of quadrat data. These functions provide a streamlined process for taking quadrat data, accounting the size of quadrats, and cleaning the data sets. Additionally, this package incorporates a Shiny app [@shiny] that uses the user's own data to produce an interactive interface where researchers can examine variables and trends in their data.
+For collaborative quadrat analyses, it is crucial that quadrat data are being properly compared to account for different methodologies. Size of quadrats varies based on the research question, study site and target species [@Benedetti:1996]; however, biomass and diversity estimates are greatly influenced by the size of quadrats [@Shmida; @williams1964; @anderson1993effect]. Thus, issues can arise when studies examine many different quadrats from various locations, which have been sampled using different quadrat sizes. If quadrat samples are being cropped to the same size before species are identified, this will produce comparable sized quadrats to be used in analysis; however, if experts have identified the species in their quadrats prior to sharing data, researchers must ensure that methods taken to standardize the size of quadrats properly considers the spatial nature of quadrat data. Currently there are no standardized methods to account for the size of quadrats and as global studies become more accessible, there is a need for standardizing these methodologies. The [quadcleanR](https://dominiquemaucieri.com/quadcleanR/) R package provides tools to assist in standardizing quadrat areas and the cleanup of quadrat data. Cleanup functions are similar to other cleanup functions, though they are tailored to quadrat analyses to provide a streamlined process for taking quadrat data, accounting the size of quadrats, and cleaning the data sets. Additionally, this package incorporates a Shiny app [@shiny] that uses the user's own data to produce an interactive interface where researchers can examine variables and trends in their data.
 
 # Cropping by area
 
