### 2022-09-06 #4674 @DominiqueMaucieri
Thank you @fboehm for your time as editor and @ViniciusBRodrigues and @diazrenata for your reviews and insight, as these comments have helped me to produce a manuscript and package that are more applicable to the wider ecological and conservation community.

Response to @diazrenata:

> Writing
> This draft of the manuscript does not contain a current state of the field section identifying how this package complements existing tools.
>
> This would be especially helpful, e.g. for the cleaning functions (a number of which are similar to functionality available in tidyverse, or operations I would be just as likely to code up from scratch specific to a given dataset) and the shiny app (which renders basic exploratory plots that seem similar to what users are probably doing on their own).
>
> (For example - a general user might find it simpler to use dplyr::select or dplyr::filter rather than learn the syntax for quadcleanR::keep_rm, or the dplyr::rename family rather than create an extra dataframe and learn quadcleanR::change_names).

Response: I have added a current state of the field section to the manuscript explaining why I believe these functions have a place amid the already immense number of cleaning functions in R. What the reviewer has said is completely true, that a comfortable coder “would be just as likely to code up from scratch” some of the functions I have made; however, this package was not created for a comfortable coder, as much as created for researchers who find themselves with quadrat data that is very uncleaned and will take a lot of coding to get their data to a cleaned state. The focus of this package is to standardize the way the quadrats are cropped for analyses, and the secondary focus is to streamline the process of cleaning quadrat data so the time of researchers can be put into the analysis of their data and not just getting the data into a usable state. Therefore, while many of these functions may resemble cleaning functions that already exist, they all have extra options and specifications which allow for these already existing functions to be built upon and accomplish many steps with a single function, especially for quadrat data.

> Documentation
> I suggest adding structure to the vignettes to tighten the focus on the added-value of the functions provided in this package. For example, the Simple Cleaning Quadrat Data vignette is a detailed narrative walk through cleaning a specific dataset, including a mix of quadcleanR and other package functions woven together. That would make it a great methods supplement to a manuscript, but it's not so easy for a user to read it and quickly understand the quadcleanR functions. There is also a fair amount of repeated text between the different vignettes, which makes it harder for a user to extract the important information from each one.
>
> You don't have to necessarily do it this way, but one suggestion would be to create a more detailed outline of the topics/functions highlighted in each vignette and structuring each section around demonstrating that group of functions.
>
> It may also be helpful to use a lightly preprocessed version of the datasets used in the vignette, to get some of the additional data cleaning out of the way so you can focus the vignette specifically on quadcleanR functions.

Response: I began with changing the language around the vignettes. These vignettes do focus on a specific kind of quadrat data, those downloaded from CoralNet, and that has been made clearer. Then to incorporate more kinds of data, I have taken the reviewers suggestion and created a very focused vignette titled “Simple Examples of Functions” to walk package users through the different clean up functions. This vignette follows the same structured outline as the cheat sheet and uses a more preprocessed dataset which I called “corals”.

> Shiny app
> The demo version of the shiny app (https://dominiquemaucieri.shinyapps.io/example/) is fun but has a lot of variables that aren't explained. With so many options, it's hard for someone who isn't familiar with this dataset to use the dashboard to generate informative plots. I think it'd be a more useful demo for a naive user if there were just a few options for faceting/grouping variables, and these were either described or given very intuitive names.

Response: I have gone back through the vignettes and added suggestions for variables to select to get a nice-looking plot. I also added another shiny example in the “Simple Examples of Functions” which is simpler and more intuitive to work with.

> Substantive scholarly effort
> I leave this up to the editor, and anything clarified in a "state-of-the-field" section of the manuscript, but from my perspective the primary unique contribution for this package is the crop_area function. There are a number of cleaning functions that are excellent utilities, but for me are in a grey area between useful scripts for streamlining a specific data cleaning workflow and general-use utilities that substantially improve on functionality already available. Again, I'm happy to see this clarified with a little more conversation.

Response: I will also leave this decision up to the editor. I believe this package will be a useful contribution to the field of ecology and I have already seen this package help collaborators cleanup their data

> Code
> Finally, I just want to note that I've run and tested the functions as they are written, but I haven't done extensive stress-testing to see if I can break them with manufactured use cases. I notice that in the source code there isn't a lot of error handling built in. I think this is reasonable for this package because most of the functions are pretty small and lightweight, but in general I would encourage more defensive programming in a package intended for widespread use.

Response: Thank you for looking into the error handling of my package. I have tried to implement error handling where I can such as when there are different lengths of vectors being used as function parameters. I have gone through and added more error handling code especially to deal with NAs and NaNs, but I would also appreciate hearing if there are ways you would add more error handling in as well.
