### 2022-10-04 #4674 @diazrenata
@DominiqueMaucieri  Thank you for the revision! It looks good to me at this point - one small point on the Shiny app, but totally optional, and more a question of streamlining the user experience!

## State of the field/statement of need

Thank you for clarifying this, and highlighting the intention of providing a standardized workflow for novices working with CoralNet data.

## Documentation

The Simple Examples of Functions vignette is a great starting place and really helps a user onboard quickly. I also appreciate including a simple version of the data, to facilitate this.

## Shiny app

I see the suggested sets of variables in the vignettes, and these are helpful.

On this point - I don't believe this is a *necessary* change, but it might be helpful for users if there is a description of variables either included or linked to from the Shiny dashboard.

## Substantive scholarly effort

Fantastic!

## Error handling

Thank you for putting some thought into this! Given that these data are coming from a relatively standardized source, this looks like an appropriate level of error handling to me.
