+++ DESCRIPTION
@@ -10,7 +10,7 @@ Description: A tool that can be customized to aid in the clean up of ecological 
 License: GPL (>= 3)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.0
+RoxygenNote: 7.2.1
 URL: https://github.com/DominiqueMaucieri/quadcleanR,
     https://dominiquemaucieri.com/quadcleanR/
 BugReports: https://github.com/DominiqueMaucieri/quadcleanR/issues
