+++ vignettes/function_examples.Rmd
@@ -64,7 +64,7 @@ corals_keep_rm <- keep_rm(corals_change_values, "DEEP" , select = "row", keep = 
 
 
 ## 4. rm_chr()
-Parts of characters can be removed based on a vector of removal characters. When workingwith images, this can be helpful to remove extra characters from image IDs, or anywhere else where you want to remove specific characters from your data. 
+Parts of characters can be removed based on a vector of removal characters. When working with images, this can be helpful to remove extra characters from image IDs, or anywhere else where you want to remove specific characters from your data. 
 
 
 ```{r rm_chr, warning=FALSE, message=FALSE}
