+++ DESCRIPTION
@@ -6,7 +6,7 @@ Authors@R:
            comment = c(ORCID = "0000-0003-1849-2472"))
 Description: A tool that can be customized to aid in the clean up of ecological data 
     collected using quadrats and can crop quadrats to ensure comparability between
-    quadrats collected under different methodologies. 
+    quadrats collected under different methodologies.
 License: GPL (>= 3)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
