+++ cran-comments.md
@@ -2,7 +2,7 @@
 Minor release after peer review with JOSS
 
 ## R CMD check results
-There are no ERRORs or WARNINGs, but there was 1 NOTE:
+There are no ERRORs or WARNINGs, but there was 2 NOTEs:
 
 #### Found only on windows-x86_64-devel:
 
@@ -15,15 +15,14 @@ Found the following files/directories:
 As noted in [R-hub issue #503](https://github.com/r-hub/rhub/issues/503), this is likely a bug/crash in MiKTeX and can likely be ignored.
 
 
-#### Found on windows-x86_64-devel, ubuntu-gcc-release and fedora-clang-devel:
+#### Found only on fedora-clang-devel:
  
 ```
-Possibly misspelled words in DESCRIPTION:
-  Quadrat (2:37)
-  quadrats (8:21, 8:43, 9:5)
+* checking HTML version of manual ... NOTE
+Skipping checking HTML validation: no command 'tidy' found
 ```
-  
-These are correctly spelt words.
+
+I cannot update Tidy on the external Fedora Linux server.
 
 ## Downstream dependencies
 There are currently no downstream dependencies for this package
