### 2022-10-15 #4674 @fboehm
@DominiqueMaucieri - I've now examined the pdf of the paper. I have one small suggestion for a fix that I ask you to implement:

1. line 46: "accounting the size" should be "accounting for the size"
Thanks again!
