### 2022-10-15 #4674 @fboehm
@DominiqueMaucieri -  A few more suggestions - in looking at the dois for the pdf, I see that the last one, for the Williams reference, seems to resolve to a book review in a journal. Can you please check on this? Does the doi point to what you want to cite?
