### 2022-10-15 #4674 @fboehm
@DominiqueMaucieri - Also, in the references, the shiny reference needs to have a capitalized "R". You can achieve this by putting curly braces around the R in your bib file, {R}. Please fix this.
