+++ manuscript/paper.bib
@@ -50,8 +50,7 @@
   Author={Williams, Carrington Bonsor},
   Year = {1964},
   Publisher={Academic Press, London},
-  Pages = {324},
-	doi = {10.2307/2257968}
+  Pages = {324}
 }
 
 
@@ -91,7 +90,7 @@
 
 
 @Manual{shiny,
-    title = {shiny: Web Application Framework for R},
+    title = {shiny: Web Application Framework for {R}},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
     year = {2021},
     note = {R package version 1.7.1},
