### 2022-10-18 #4674 @DominiqueMaucieri
@fboehm  The new version is v1.1.1 and I have archived it on Zenodo with DOI: 10.5281/zenodo.7222742
